+++ source/extensions/omni.flux.utils.common/omni/flux/utils/common/symlink.py
@@ -17,14 +17,13 @@
 
 import subprocess
 import sys
-from typing import Tuple
 
 from .uac import UnsupportedPlatformError as _UnsupportedPlatformError
 from .uac import is_admin as _is_admin
 from .uac import sudo as _sudo
 
 
-def create_folder_symlinks(links_targets: list[Tuple[str, str]], create_junction: bool = False):
+def create_folder_symlinks(links_targets: list[tuple[str, str]], create_junction: bool = False):
     """
     Create symlink(s). If create_junction is False and the user doesn't have the permission to create symlink(s),
     it will prompt the Windows UAC for Windows user.