+++ source/extensions/omni.flux.utils.material_converter/omni/flux/utils/material_converter/base/attribute_base.py
@@ -15,7 +15,8 @@
 * limitations under the License.
 """
 
-from typing import Any, Callable, Optional, Tuple
+from collections.abc import Callable
+from typing import Any
 
 from pxr import Sdf, Usd
 from pydantic import BaseModel, Field
@@ -37,7 +38,7 @@ def _translate(input_attr_value: Any, input_attr: Usd.Attribute) -> Any:
 
 def _translate_alt(
     input_attr_type: Sdf.ValueTypeNames, input_attr_value: Any, input_attr: Usd.Attribute
-) -> Tuple[Sdf.ValueTypeNames, Any]:
+) -> tuple[Sdf.ValueTypeNames, Any]:
     """
     Allows translating the input type and value to produce the output type and value when we need to create
     attributes in the output material.
@@ -61,7 +62,7 @@ class AttributeBase(BaseModel):
     # Attribute name used on the output shader
     output_attr_name: str
     # If set, this value will be used to create the attribute on the output shader when it doesn't exist on the input
-    output_default_value: Optional[Any] = None
+    output_default_value: Any | None = None
     # Function used to translate the value of the input shader to the value of the output shader
     translate_fn: Callable[[Any, Usd.Attribute], Any] = Field(default=_translate)
     # tell if the attribute is a real attribute that exists by default, or if this is a fake one that was created
@@ -69,5 +70,5 @@ class AttributeBase(BaseModel):
     # Function used to translate the value of the input shader to the value of the output shader when we need to
     # create the attribute in the output material first.
     translate_alt_fn: Callable[
-        [Optional[Sdf.ValueTypeNames], Any, Optional[Usd.Attribute]], Tuple[Optional[Sdf.ValueTypeNames], Any]
+        [Sdf.ValueTypeNames | None, Any, Usd.Attribute | None], tuple[Sdf.ValueTypeNames | None, Any]
     ] = Field(default=_translate_alt)