+++ source/extensions/omni.flux.utils.material_converter/omni/flux/utils/material_converter/impl/omni_pbr_to_aperture_pbr.py
@@ -18,7 +18,6 @@
 from __future__ import annotations
 
 from enum import Enum
-from typing import Optional, Tuple
 
 from omni.flux.utils.material_converter.base.attribute_base import AttributeBase
 from omni.flux.utils.material_converter.base.converter_base import ConverterBase
@@ -84,8 +83,8 @@ class OmniPBRToAperturePBRConverterBuilder(ConverterBuilderBase):
         return _NormalMapEncodings.TANGENT_SPACE_DX.value if value else _NormalMapEncodings.TANGENT_SPACE_OGL.value
 
     def _convert_normal_encoding_alt(
-        self, _: Sdf.ValueTypeNames, value: bool, input_attr: Optional[Usd.Attribute]
-    ) -> Tuple[Sdf.ValueTypeNames, int]:
+        self, _: Sdf.ValueTypeNames, value: bool, input_attr: Usd.Attribute | None
+    ) -> tuple[Sdf.ValueTypeNames, int]:
         """Alternate translate method to create attributes in the output material"""
         return (
             Sdf.ValueTypeNames.Int,